Make lipstick wait until ofono is ready
This is a bit hacky but read on.
Currently, ofono assumes the sim slot / imei is known immediately.  With
the pinephone, ofono and gobi plugin, the imei is signalled after
startup.  If lipstick is started before the imei/slots are known, then
it assumes no mobile network, unless it is restarted.

This adds 2 service changes.
1. Make ofono wait 30 seconds, and then write a file to indicate it has
   finished startup
2. Make lipstick wait until this file has been created before starting.

It is done this way becuase ofono is a system service, and lipstick is a
user-session service, so making them depend on eachother wasnt working.